requirements.txt
# external requirements
click
Sphinx
coverage
awscli>=1.27.130
flake8
python-dotenv>=0.5.1
requests
requests-mock
scipy
geosyspy
pandas
shapely
rasterio
xarray
pyproj
matplotlib
fastapi
hypercorn[trio]==0.14.3